refactor: add new resolver package
