+++ internal/auth/resolver/formresolver/path_resolver.go
@@ -1,4 +1,4 @@
-package form
+package formresolver
 
 import (
 	"NYCU-SDC/core-system-backend/internal"
