+++ internal/auth/resolver/unitresolver/path_resolver.go
@@ -1,4 +1,4 @@
-package unit
+package unitresolver
 
 import (
 	"NYCU-SDC/core-system-backend/internal"
